refactor hyperparameter tuning into its own module
