+++ src/microcalibrate/__init__.py
@@ -1,2 +1,7 @@
 from .calibration import Calibration
-from .evaluation import evaluate_estimate_distance_to_targets
+from .evaluation import (
+    evaluate_estimate_distance_to_targets,
+    evaluate_holdout_robustness,
+    evaluate_sparse_weights,
+)
+from .hyperparameter_tuning import tune_l0_hyperparameters
